Move the macros in a module
